+++ src/datasources/alarms/AlarmsDataSource.test.ts
@@ -4,7 +4,7 @@ import { setupDataSource, requestMatching, createFetchResponse, createFetchError
 import { AlarmsDataSource } from './AlarmsDataSource';
 import { DataQueryRequest } from '@grafana/data';
 import { QueryType, AlarmsVariableQuery } from './types/types';
-import { AlarmsCountDataSource } from './query-type-handlers/alarms-count/AlarmsCountDataSource';
+import { AlarmsCountQueryHandler } from './query-type-handlers/alarms-count/AlarmsCountQueryHandler';
 import { QUERY_ALARMS_RELATIVE_PATH } from './constants/QueryAlarms.constants';
 
 let datastore: AlarmsDataSource, backendServer: MockProxy<BackendSrv>;
@@ -17,36 +17,36 @@ describe('AlarmsDataSource', () => {
   });
 
   it('should initialize with AlarmsCount as the default query', () => {
-    expect(datastore.defaultQuery).toEqual({ queryType: QueryType.AlarmsCount });
+    expect(datastore.defaultQuery).toEqual({ queryType: QueryType.AlarmsCount, filter: '' });
   });
 
-  describe('AlarmsCountDataSource', () => {
+  describe('AlarmsCountQueryHandler', () => {
     const query = { refId: 'A', queryType: QueryType.AlarmsCount };
 
-    let alarmsCountDataSource: AlarmsCountDataSource;
+    let alarmsCountQueryHandler: AlarmsCountQueryHandler;
     
     beforeEach(() => {
-      alarmsCountDataSource = datastore.alarmsCountDataSource;
+      alarmsCountQueryHandler = datastore.alarmsCountQueryHandler;
     });
 
     describe('runQuery', () => {
-      it('should call AlarmsCountDataSource runQuery when queryType is AlarmsCount', async () => {
-        alarmsCountDataSource.runQuery = jest.fn().mockResolvedValue({ refId: "A", fields: [] });
+      it('should call AlarmsCountQueryHandler runQuery when queryType is AlarmsCount', async () => {
+        alarmsCountQueryHandler.runQuery = jest.fn().mockResolvedValue({ refId: 'A', fields: [] });
 
         const result = await datastore.runQuery(query, dataQueryRequest);
 
-        expect(alarmsCountDataSource.runQuery).toHaveBeenCalledWith(query, dataQueryRequest);
-        expect(result).toEqual({ refId: "A", fields: [] });
+        expect(alarmsCountQueryHandler.runQuery).toHaveBeenCalledWith(query, dataQueryRequest);
+        expect(result).toEqual({ refId: 'A', fields: [] });
       });
     });
 
     describe('shouldRunQuery', () => {
-      it('should call AlarmsCountDataSource shouldRunQuery when queryType is AlarmsCount', () => {
-        alarmsCountDataSource.shouldRunQuery = jest.fn().mockReturnValue(true);
+      it('should call AlarmsCountQueryHandler shouldRunQuery when queryType is AlarmsCount', () => {
+        alarmsCountQueryHandler.shouldRunQuery = jest.fn().mockReturnValue(true);
 
         const result = datastore.shouldRunQuery(query);
 
-        expect(alarmsCountDataSource.shouldRunQuery).toHaveBeenCalled();
+        expect(alarmsCountQueryHandler.shouldRunQuery).toHaveBeenCalled();
         expect(result).toBe(true);
       });
     });