+++ src/datasources/alarms/AlarmsDataSource.ts
@@ -2,14 +2,14 @@ import { DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, TestDataSou
 import { BackendSrv, TemplateSrv, getBackendSrv, getTemplateSrv } from '@grafana/runtime';
 import { DataSourceBase } from 'core/DataSourceBase';
 import { AlarmsQuery, AlarmsVariableQuery, QueryType } from './types/types';
-import { AlarmsCountDataSource } from './query-type-handlers/alarms-count/AlarmsCountDataSource';
+import { AlarmsCountQueryHandler } from './query-type-handlers/alarms-count/AlarmsCountQueryHandler';
 import { QUERY_ALARMS_RELATIVE_PATH } from './constants/QueryAlarms.constants';
 import { ListAlarmsDataSource } from './query-type-handlers/list-alarms/ListAlarmsDataSource';
 
 export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
   public readonly defaultQuery: Omit<AlarmsQuery, 'refId'>;
 
-  private readonly _alarmsCountDataSource: AlarmsCountDataSource;
+  private readonly _alarmsCountQueryHandler: AlarmsCountQueryHandler;
   private readonly _listAlarmsDataSource: ListAlarmsDataSource;
 
   constructor(
@@ -18,17 +18,17 @@ export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
     readonly templateSrv: TemplateSrv = getTemplateSrv()
   ) {
     super(instanceSettings, backendSrv, templateSrv);
-    this._alarmsCountDataSource = new AlarmsCountDataSource(instanceSettings, backendSrv, templateSrv);
+    this._alarmsCountQueryHandler = new AlarmsCountQueryHandler(instanceSettings, backendSrv, templateSrv);
     this._listAlarmsDataSource = new ListAlarmsDataSource(instanceSettings, backendSrv, templateSrv);
 
     // AB#3064461 - Update defaultQuery to use list alarms defaults when supported
-    this.defaultQuery = this._alarmsCountDataSource.defaultQuery;
+    this.defaultQuery = this._alarmsCountQueryHandler.defaultQuery;
   }
 
-  public async runQuery(query: AlarmsQuery, _: DataQueryRequest): Promise<DataFrameDTO> {
+  public async runQuery(query: AlarmsQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
     switch (query.queryType) {
       case QueryType.AlarmsCount:
-        return this.alarmsCountDataSource.runQuery(query, _);
+        return this.alarmsCountQueryHandler.runQuery(query, options);
       default:
         throw new Error('Invalid query type');
     }
@@ -37,14 +37,14 @@ export class AlarmsDataSource extends DataSourceBase<AlarmsQuery> {
   public shouldRunQuery(query: AlarmsQuery): boolean {
     switch (query.queryType) {
       case QueryType.AlarmsCount:
-        return this.alarmsCountDataSource.shouldRunQuery(query);
+        return this.alarmsCountQueryHandler.shouldRunQuery(query);
       default:
         return false;
     }
   }
 
-  public get alarmsCountDataSource(): AlarmsCountDataSource {
-    return this._alarmsCountDataSource;
+  public get alarmsCountQueryHandler(): AlarmsCountQueryHandler {
+    return this._alarmsCountQueryHandler;
   }
 
   public get listAlarmsDataSource(): ListAlarmsDataSource {