+++ src/datasources/alarms/constants/DefaultQueries.contants.ts
@@ -1,0 +1,7 @@
+import { AlarmsCountQuery } from '../types/AlarmsCount.types';
+import { QueryType } from '../types/types';
+
+export const defaultAlarmsCountQuery: Omit<AlarmsCountQuery, 'refId'> = {
+    queryType: QueryType.AlarmsCount,
+    filter: '',
+};