+++ src/datasources/data-frame/DataFrameDataSourceBase.test.ts
@@ -2,6 +2,16 @@ import { DataFrameDTO, DataQueryRequest, DataSourceInstanceSettings, TimeRange }
 import { BackendSrv, TemplateSrv } from '@grafana/runtime';
 import { DataFrameDataSourceBase } from './DataFrameDataSourceBase';
 import { DataFrameQuery, DataFrameDataSourceOptions, TableProperties, TableDataRows, Column } from './types';
+import { WorkspaceUtils } from 'shared/workspace.utils';
+import { Workspace } from 'core/types';
+
+jest.mock('core/utils', () => ({
+    ...jest.requireActual('core/utils'),
+    getVariableOptions: jest.fn(() => [
+        { label: 'Var1', value: 'Value1' },
+        { label: 'Var2', value: 'Value2' },
+    ]),
+}));
 
 describe('DataFrameDataSourceBase', () => {
     let instanceSettings: DataSourceInstanceSettings<DataFrameDataSourceOptions>;
@@ -97,5 +107,100 @@ describe('DataFrameDataSourceBase', () => {
         await expect(ds.getTableProperties()).resolves.toEqual({});
         await expect(ds.getDecimatedTableData({} as DataFrameQuery, [], {} as TimeRange)).resolves.toEqual([]);
         await expect(ds.queryTables('')).resolves.toEqual([]);
+    });
+
+    it('should return global variable options when the `globalVariableOptions` is called', async () => {
+        const ds = new TestDataFrameDataSource(instanceSettings, backendSrv, templateSrv);
+        const options = await ds.globalVariableOptions();
+
+        expect(options).toEqual([
+            { label: 'Var1', value: 'Value1' },
+            { label: 'Var2', value: 'Value2' },
+        ]);
+    });
+
+    describe('loadWorkspaces', () => {
+        let ds: DataFrameDataSourceBase;
+        let getWorkspacesSpy: jest.SpyInstance;
+
+        beforeEach(() => {
+            ds = new TestDataFrameDataSource(instanceSettings, backendSrv, templateSrv);
+            getWorkspacesSpy = jest.spyOn(WorkspaceUtils.prototype, 'getWorkspaces');
+        });
+
+        it('should return workspaces', async () => {
+            getWorkspacesSpy.mockResolvedValue(
+                new Map([
+                    ['1', { id: '1', name: 'WorkspaceName' } as Workspace],
+                    ['2', { id: '2', name: 'AnotherWorkspaceName' } as Workspace],
+                ])
+            );
+
+            const result = await ds.loadWorkspaces();
+
+            expect(result.get('1')?.name).toBe('WorkspaceName');
+            expect(result.get('2')?.name).toBe('AnotherWorkspaceName');
+        });
+
+        it('should handle errors and set error and innerError fields', async () => {
+            getWorkspacesSpy.mockRejectedValue(new Error('Error'));
+
+            await ds.loadWorkspaces();
+
+            expect(ds.errorTitle).toBe('Warning during dataframe query');
+            expect(ds.errorDescription).toContain(
+                'Some values may not be available in the query builder lookups due to an unknown error.'
+            );
+        });
+
+        it('should handle errors and set innerError fields with error message detail', async () => {
+            ds.errorTitle = '';
+            getWorkspacesSpy.mockRejectedValue(
+                new Error('Request failed with status code: 500, Error message: {"message": "Internal Server Error"}')
+            );
+
+            await ds.loadWorkspaces();
+
+            expect(ds.errorTitle).toBe('Warning during dataframe query');
+            expect(ds.errorDescription).toContain(
+                'Some values may not be available in the query builder lookups due to the following error: Internal Server Error.'
+            );
+        });
+
+        it('should throw timeOut error when API returns 504 status', async () => {
+            ds.errorTitle = '';
+            getWorkspacesSpy.mockRejectedValue(new Error('Request failed with status code: 504'));
+
+            await ds.loadWorkspaces();
+
+            expect(ds.errorTitle).toBe('Warning during dataframe query');
+            expect(ds.errorDescription).toContain(
+                `The query builder lookups experienced a timeout error. Some values might not be available. Narrow your query with a more specific filter and try again.`
+            );
+        });
+
+        it('should throw too many requests error when API returns 429 status', async () => {
+            ds.errorTitle = '';
+            getWorkspacesSpy.mockRejectedValue(new Error('Request failed with status code: 429'));
+
+            await ds.loadWorkspaces();
+
+            expect(ds.errorTitle).toBe('Warning during dataframe query');
+            expect(ds.errorDescription).toContain(
+                `The query builder lookups failed due to too many requests. Please try again later.`
+            );
+        });
+
+        it('should throw not found error when API returns 404 status', async () => {
+            ds.errorTitle = '';
+            getWorkspacesSpy.mockRejectedValue(new Error('Request failed with status code: 404'));
+
+            await ds.loadWorkspaces();
+
+            expect(ds.errorTitle).toBe('Warning during dataframe query');
+            expect(ds.errorDescription).toContain(
+                `The query builder lookups failed because the requested resource was not found. Please check the query parameters and try again.`
+            );
+        });
     });
 });